+++ ui/src/main/java/io/snabble/sdk/ui/checkout/CheckoutView.java
@@ -97,9 +97,7 @@ public class CheckoutView extends FrameLayout implements Checkout.OnCheckoutStat
                 displayPaymentView();
                 break;
             case PAYMENT_APPROVED:
-            case RECEIPT_AVAILABLE:
-                if (currentState == Checkout.State.PAYMENT_APPROVED
-                 || currentState == Checkout.State.RECEIPT_AVAILABLE) {
+                if (currentState == Checkout.State.PAYMENT_APPROVED) {
                     break;
                 }
 
@@ -132,7 +130,6 @@ public class CheckoutView extends FrameLayout implements Checkout.OnCheckoutStat
     }
 
     private void displayPaymentView() {
-        Project project = SnabbleUI.getProject();
         switch (checkout.getSelectedPaymentMethod()) {
             case DE_DIRECT_DEBIT:
                 displayView(new CheckoutStatusView(getContext()));
@@ -142,46 +139,8 @@ public class CheckoutView extends FrameLayout implements Checkout.OnCheckoutStat
                 checkoutQRCodePOSView.setQRCodeText(checkout.getQRCodePOSContent());
                 displayView(checkoutQRCodePOSView);
                 break;
-            case ENCODED_CODES:
+            case QRCODE_OFFLINE:
                 displayView(new CheckoutEncodedCodesView(getContext()));
-                break;
-            case ENCODED_CODES_CSV:
-                displayView(new CheckoutEncodedCodesView(getContext(),
-                        new EncodedCodesOptions.Builder(project)
-                        .prefix("snabble;\n")
-                        .separator("\n")
-                        .suffix("")
-                        .repeatCodes(false)
-                        .countSeparator(";")
-                        .maxCodes(100)
-                        .build()));
-                break;
-            case ENCODED_CODES_IKEA:
-                EncodedCodesOptions options = project.getEncodedCodesOptions();
-                int maxCodes = 45;
-                int maxChars = EncodedCodesOptions.DEFAULT_MAX_CHARS;
-                if (options != null) {
-                    maxCodes = options.maxCodes;
-                    maxChars = options.maxChars;
-                }
-
-                String prefix = "9100003\u001d100{qrCodeCount}\u001d240";
-
-                String prefixWithCustomerCard = "9100003\u001d100{qrCodeCount}";
-                if (project.getCustomerCardId() != null) {
-                    prefixWithCustomerCard += "\u001d92" + project.getCustomerCardId();
-                }
-                prefixWithCustomerCard += "\u001d240";
-
-                displayView(new CheckoutEncodedCodesView(getContext(),
-                        new EncodedCodesOptions.Builder(project)
-                        .prefix(prefix)
-                        .prefix(0, prefixWithCustomerCard)
-                        .separator("\u001d240")
-                        .suffix("")
-                        .maxCodes(maxCodes)
-                        .maxChars(maxChars)
-                        .build()));
                 break;
         }
     }